requirements.txt
streamlit
requests
python-dotenv
groq
pandas
numpy
psycopg2-binary
sqlalchemy
openai
pymupdf
pytesseract
Pillow
